refactor(utils): split questionnaire-api-converters under the line cap (#544)
Extraction only, zero behavior change. Splits by concern into three
co-located modules: questionnaire-api-converters-items.ts (API -> Local
item converters + their loose input shapes) and
questionnaire-api-converters-init.ts (initializeFromApiData, the
timedelta parsers, and normalizeQuestionCollections). The original
module keeps the Local -> API update converters and re-exports every
prior symbol so importers and tests need no changes.
